Test: for single float

+++ lib/kata_3.rb
@@ -1,3 +1,3 @@
 def to_number_array(string_array)
-  string_array.map{|x| x.to_i}
+  string_array.map{|x| x.to_f}
 end
